Level 1 and Level 2 Quiz Preparation/01 OpenAI Agents SDK_Agent class/main.py
from agents import Agent, Runner
# , set_tracing_disabled
from dotenv import load_dotenv

load_dotenv()
# set_tracing_disabled(True)

agent = Agent(
    name="Assistant",
    instructions="you are helpful assistant",
)

result = Runner.run_sync(
    agent,
    "Hello"
)

print(result.final_output)