Level 1 and Level 2 Quiz Preparation/01 OpenAI Agents SDK_Agent class/pyproject.toml
[project]
name = "01-openai-agents-sdk-agent-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
]
